refactor(sre-copilot): single-file student structure (edit only solution.py)
Restructure the agent so students touch ONE file and enable modules by
uncommenting lines (beginner-friendly). Removes the MODULE env toggle.

- solution.py: the only student-edited file — build_agent(session_id) with
  commented module lines (M2 tools / M3 memory / M4 KB / M5 AWS docs MCP).
- coffee_sre/ framework package (hidden): prompt, model, tools.ALL,
  knowledge_base.search, aws_docs.load() (MCP singleton), memory(session_id),
  runtime (FastAPI/SSE, imports solution.build_agent).
- Delete agent/ package, steps/run_local.py, infra/set_module.sh.
- CFN: drop StartModule/MODULE; Handler run.sh -> uvicorn coffee_sre.runtime:app.
- packager bundles coffee_sre/ + solution.py.
- README rewritten around the single-file uncomment flow.

Verified deployed: M1 (all commented) runs, no tools, asks operator for evidence.
