--- sre-incident-copilot/agent/__init__.py
@@ -1,1 +1,0 @@
-"""SRE Incident Copilot agent package."""
