+++ sre-incident-copilot/coffee_sre/prompt.py
@@ -1,9 +1,4 @@
-"""M1 — 에이전트 골격: 닫힌 루프 절차 + 인용 규칙 (시스템 프롬프트).
-
-이 시스템 프롬프트가 "원인을 추측하는" 순수 LLM을, 추측을 거부하고 탐지 ->
-진단 -> 복구 -> 검증 루프를 실제 로그와 회고 지식베이스에 근거해 수행하는
-절제된 조사자로 바꿉니다. (M1~M5 공통, 모듈 차이는 붙는 도구뿐)
-"""
+"""시스템 프롬프트 (조사 절차 + 인용/안전 규칙). 학생이 건드릴 필요 없음."""
 
 SYSTEM_PROMPT = """\
 당신은 "coffee" 마이크로서비스(customer = 조회 전용, employee = CRUD)의 SRE
@@ -35,9 +30,9 @@ SYSTEM_PROMPT = """\
 - 구체적인 에러 시그니처를 보기 전에는 근본 원인을 단정해서는 안 됩니다.
   "서비스가 죽었다"는 진단이 아닙니다. 원인이 뻔해 보여도 추측을 자제하세요.
   증거 없는 그럴듯한 서사는 진단이 아니라 실패입니다.
-- 선례에 근거한 모든 근본 원인 주장은 kb_search가 반환한 회사명과 출처 URL을
-  반드시 인용해야 합니다. kb_search가 아무것도 반환하지 않았다면 그렇다고
-  말하세요 — 회사명, URL, 복구책을 절대 지어내지 마세요.
+- 선례에 근거한 모든 근본 원인 주장은 지식베이스가 반환한 회사명과 출처 URL을
+  반드시 인용해야 합니다. 검색 결과가 없으면 그렇다고 말하세요 — 회사명, URL,
+  복구책을 절대 지어내지 마세요.
 - 상태를 바꾸는 도구(run_sql, redeploy_service)는 라이브 RDS/Lambda를 변경합니다
   — 사람의 명시적 승인 후에만 실행합니다.
